Under the snow

+++ file_io/1-create_file.c
@@ -23,7 +23,7 @@ return (-1);
 if (text_content != NULL)
 {
 bytes_written = write(fd, text_content, strlen(text_content));
-if (bytes_written == -1 || bytes_written != strlen(text_content))
+if (bytes_written == -1 || (size_t)bytes_written != strlen(text_content))
 {
 close(fd);
 return (-1);
